+++ drools-core/src/main/java/org/drools/core/base/accumulators/JavaAccumulatorFunctionExecutor.java
@@ -16,6 +16,14 @@
 
 package org.drools.core.base.accumulators;
 
+import java.io.Externalizable;
+import java.io.IOException;
+import java.io.ObjectInput;
+import java.io.ObjectOutput;
+import java.io.Serializable;
+import java.util.HashMap;
+import java.util.Map;
+
 import org.drools.core.WorkingMemory;
 import org.drools.core.common.InternalFactHandle;
 import org.drools.core.rule.Declaration;
@@ -26,14 +34,6 @@ import org.drools.core.spi.ReturnValueExpression.SafeReturnValueExpression;
 import org.drools.core.spi.Tuple;
 import org.drools.core.spi.Wireable;
 import org.kie.internal.security.KiePolicyHelper;
-
-import java.io.Externalizable;
-import java.io.IOException;
-import java.io.ObjectInput;
-import java.io.ObjectOutput;
-import java.io.Serializable;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * A Java accumulator function executor implementation
@@ -80,7 +80,7 @@ public class JavaAccumulatorFunctionExecutor
         JavaAccumulatorFunctionContext context = new JavaAccumulatorFunctionContext();
         context.context = this.function.createContext();
         if ( this.function.supportsReverse() ) {
-            context.reverseSupport = new HashMap<Integer, Object>();
+            context.reverseSupport = new HashMap<Long, Object>();
         }
         return context;
     }
@@ -169,7 +169,7 @@ public class JavaAccumulatorFunctionExecutor
         implements
         Externalizable {
         public Serializable               context;
-        public Map<Integer, Object>       reverseSupport;
+        public Map<Long, Object>          reverseSupport;
 
         public JavaAccumulatorFunctionContext() {
         }
@@ -178,7 +178,7 @@ public class JavaAccumulatorFunctionExecutor
         public void readExternal(ObjectInput in) throws IOException,
                                                 ClassNotFoundException {
             context = (Externalizable) in.readObject();
-            reverseSupport = (Map<Integer, Object>) in.readObject();
+            reverseSupport = (Map<Long, Object>) in.readObject();
         }
 
         public void writeExternal(ObjectOutput out) throws IOException {
