+++ drools-core/src/main/java/org/drools/core/common/QueryElementFactHandle.java
@@ -16,33 +16,33 @@
 
 package org.drools.core.common;
 
-import org.drools.core.factmodel.traits.TraitTypeEnum;
-import org.drools.core.reteoo.LeftTuple;
-import org.drools.core.reteoo.RightTuple;
-import org.drools.core.spi.Tuple;
+import java.util.Arrays;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
-import java.util.Arrays;
+import org.drools.core.factmodel.traits.TraitTypeEnum;
+import org.drools.core.reteoo.LeftTuple;
+import org.drools.core.reteoo.RightTuple;
+import org.drools.core.spi.Tuple;
 
 @XmlAccessorType(XmlAccessType.NONE)
 public class QueryElementFactHandle
     implements
     InternalFactHandle {
     private Object object;
-    private int id;
+    private long id;
     private int identityHashCode;
     private long recency;
     private boolean                 negated;
 
     protected QueryElementFactHandle() {}
 
-    public QueryElementFactHandle(Object object, int id, long recency) {
+    public QueryElementFactHandle(Object object, long id, long recency) {
         this( object, id, DefaultFactHandle.determineIdentityHashCode( object ), recency );
     }
 
-    public QueryElementFactHandle(Object object, int id, int identityHashCode, long recency) {
+    public QueryElementFactHandle(Object object, long id, int identityHashCode, long recency) {
         this.object = object;
         this.id = id;
         this.recency = recency;
@@ -59,7 +59,7 @@ public class QueryElementFactHandle
         this.negated = negated;
     }
 
-    public int getId() {
+    public long getId() {
         return this.id;
     }
