+++ drools-core/src/main/java/org/drools/core/reteoo/LeftTupleImpl.java
@@ -16,10 +16,10 @@
 
 package org.drools.core.reteoo;
 
+import java.util.Arrays;
+
 import org.drools.core.common.InternalFactHandle;
 import org.drools.core.spi.PropagationContext;
-
-import java.util.Arrays;
 
 public class LeftTupleImpl extends BaseLeftTuple {
     private static final long serialVersionUID = 540l;
@@ -159,7 +159,7 @@ public class LeftTupleImpl extends BaseLeftTuple {
     protected String toExternalString() {
         StringBuilder builder = new StringBuilder();
         builder.append(String.format("%08X", System.identityHashCode(this))).append(":");
-        int[] ids = new int[getIndex() + 1];
+        long[] ids = new long[getIndex() + 1];
         LeftTuple entry = this;
         while (entry != null) {
             if ( entry.getFactHandle() != null ) {
